Rename regex_new to regex
