+++ lib/engine_rules/js.js
@@ -1,7 +1,7 @@
 var _ = require('lodash');
 var sub = require('string-sub');
 
-var REGEX = require('../regex_new');
+var REGEX = require('../regex');
 
 var getRegExp = REGEX.getRegExp;
