+++ lib/formatters/html.js
@@ -2,7 +2,7 @@ var _ = require('lodash');
 var Formatter = require('content-formatter');
 var sub = require('string-sub');
 
-var REGEX = require('../regex_new');
+var REGEX = require('../regex');
 
 var getRegExp = REGEX.getRegExp;
