+++ lib/regex.js
